Landing Page_Body Refactor
