+++ SDCTA/src/App.css
@@ -25,76 +25,6 @@
 	pointer-events: none;
 }
 
-/* Navbar Styles */
-.navbar {
-	display: flex;
-	align-items: center;
-	height: 106px;
-	background-color: white;
-	font-weight: 600;
-
-}
-
-.nav-items {
-	display: flex;
-	justify-content: space-between;
-	width: 40%;
-	margin-left: 60px;
-}
-
-.nav-item {
-	color: black;
-	text-decoration: none;
-}
-
-.active-page {
-	color: #7F1922;
-}
-
-.nav-item:hover {
-	color: red;
-}
-
-.nav-actions {
-	display: flex;
-
-	justify-content: flex-end;
-
-	width: 50%;
-}
-
-.login {
-	color: black;
-	text-decoration: none;
-	margin-right: 20px;
-	padding: 15px
-}
-
-.login:hover {
-	color: red;
-}
-
-.sign-up {
-	color: white;
-	text-decoration: none;
-	margin-right: 20px;
-	background-color: red;
-	padding: 15px;
-	border-radius: 4px
-}
-
-.sign-up:hover {
-	background-color: white;
-	color: red;
-}
-
-.logo {
-	width: 174px;
-	height: 80px;
-	padding-left: 20px;
-
-}
-
 @media (prefers-reduced-motion: no-preference) {
 	.App-logo {
 		animation: App-logo-spin infinite 20s linear;
