+++ SDCTA/src/pages/LandingPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Body } from '../components/Body';
+import { Body } from '../components/Landing Page/Body';
 import {Footer} from '../components/Footer';
 
 export const LandingPage = () => {
